fixes #8366 - add parameter type to unique index

+++ db/migrate/20140805114754_add_unique_index_to_parameter.rb
@@ -1,9 +1,11 @@
 class AddUniqueIndexToParameter < ActiveRecord::Migration
   def up
-    add_index :parameters, [:reference_id, :name], :unique => true
+    add_index :parameters, [:type, :reference_id, :name], :unique => true
   end
 
   def down
-    remove_index :parameters, :column => [:reference_id, :name]
+    # previous version, prior to #8366 and 20141112165600_add_type_to_parameter_index
+    remove_index :parameters, :column => [:reference_id, :name] if index_exists? :parameters, [:reference_id, :name], :unique => true
+    remove_index :parameters, :column => [:type, :reference_id, :name] if index_exists? :parameters, [:type, :reference_id, :name], :unique => true
   end
 end
